change view layout
