+++ base_custom_attributes/custom_attributes_view.xml
@@ -10,11 +10,15 @@
             <field name="model">attribute.set</field>
             <field name="arch" type="xml">
                 <form string="Attribute Set">
-                    <field name="name"/>
-                    <field name="model_id" invisible="context.get('force_model')"/>
-                    <field name="attribute_group_ids" colspan="4"
-                           context="{'default_model_id': model_id, 'from_attribute_set':True}">
-                    </field>
+                    <sheet>
+                        <group>
+                            <field name="name"/>
+                            <field name="model_id" invisible="context.get('force_model')"/>
+                            <field name="attribute_group_ids" colspan="4"
+                                   context="{'default_model_id': model_id, 'from_attribute_set':True}">
+                            </field>
+                        </group>
+                    </sheet>
                 </form>
             </field>
         </record>
@@ -55,26 +59,30 @@
              action="attribute_set_form_action" id="menu_attribute_set_action"
              parent="menu_attribute_in_admin" sequence="10"/>
  
-    <!-- ATTRIBUTE GROUP VIEW -->
+        <!-- ATTRIBUTE GROUP VIEW -->
         <record id="attribute_group_form_view" model="ir.ui.view">
             <field name="name">attribute.group.form</field>
             <field name="model">attribute.group</field>
             <field name="arch" type="xml">
                 <form string="Attribute Group">
-                    <field name="name" />
-                    <field name="sequence" />
-                    <field name="model_id" invisible="context.get('from_attribute_set') or context.get('force_model')"/>
-                    <field name="attribute_set_id" invisible="context.get('from_attribute_set')"/>
-                    <field name="attribute_ids" colspan="4" nolabel="1">
-                        <form string="Attribute Location">
-                            <field name="attribute_id" context="{'default_model_id': parent.model_id}"/>
+                    <sheet>
+                        <group>
+                            <field name="name" />
                             <field name="sequence" />
-                        </form>
-                        <tree string="Attribute Location" editable="top">
-                            <field name="attribute_id" context="{'default_model_id': parent.model_id}"/>
-                            <field name="sequence" />
-                        </tree>
-                    </field>
+                            <field name="model_id" invisible="context.get('from_attribute_set') or context.get('force_model')"/>
+                            <field name="attribute_set_id" invisible="context.get('from_attribute_set')"/>
+                            <field name="attribute_ids" colspan="4" nolabel="1">
+                                <form string="Attribute Location">
+                                    <field name="attribute_id" context="{'default_model_id': parent.model_id}"/>
+                                    <field name="sequence" />
+                                </form>
+                                <tree string="Attribute Location" editable="top">
+                                    <field name="attribute_id" context="{'default_model_id': parent.model_id}"/>
+                                    <field name="sequence" />
+                                </tree>
+                            </field>
+                        </group>
+                    </sheet>
                 </form>
             </field>
         </record>
@@ -118,21 +126,23 @@
              action="attribute_group_form_action" id="menu_attribute_group_action"
              parent="menu_attribute_in_admin" sequence="20"/>
 
-    <!-- ATTRIBUTE VIEW -->
-
+        <!-- ATTRIBUTE VIEW -->
         <record id="attribute_attribute_form_view" model="ir.ui.view">
             <field name="name">attribute.attribute.form</field>
             <field name="model">attribute.attribute</field>
             <field name="arch" type="xml">
                 <form string="Attribute">
-                    <field name="field_description" />
-                    <field name="name" />
-                    <field name="attribute_type" />
-                    <field name="model_id" />
-                    <field name="serialized" attrs="{'invisible':[('attribute_type', '=', 'multiselect')]}"/>
-                    <field name="size" attrs="{'invisible':[('attribute_type', '!=', 'char')]}"/>
-                    <field name="required_on_views"/>
-                    <field name="translate" attrs="{'invisible':[('attribute_type', 'not in', ('char', 'text'))]}"/>
+                    <sheet>
+                    <group>
+                        <field name="field_description" />
+                        <field name="name" />
+                        <field name="attribute_type" />
+                        <field name="model_id" />
+                        <field name="serialized" attrs="{'invisible':[('attribute_type', '=', 'multiselect')]}"/>
+                        <field name="size" attrs="{'invisible':[('attribute_type', '!=', 'char')]}"/>
+                        <field name="required_on_views"/>
+                        <field name="translate" attrs="{'invisible':[('attribute_type', 'not in', ('char', 'text'))]}"/>
+                    </group>  
                     <newline />
                     <group colspan="4" attrs="{'invisible':[('attribute_type', 'not in', ['select', 'multiselect'])]}">
                         <group groups="base.group_advanced_attribute">
@@ -148,6 +158,7 @@
                         </field>
                     </group>
                     <field name="create_date" invisible="1"/>
+                    </sheet>
                 </form>
             </field>
         </record>
@@ -186,7 +197,7 @@
             action="attribute_attribute_form_action" id="menu_attribute_attribute_action"
             parent="menu_attribute_in_admin" sequence="30"/> -->
 
-    <!-- ATTRIBUTE OPTION VIEW -->
+        <!-- ATTRIBUTE OPTION VIEW -->
         <record id="attribute_option_form_popup_view" model="ir.ui.view">
             <field name="name">attribute.option.form.popup</field>
             <field name="model">attribute.option</field>
@@ -205,10 +216,14 @@
             <field name="model">attribute.option</field>
             <field name="arch" type="xml">
                 <form string="Attribute Option" col="6">
-                    <field name="name" colspan="2"/>
-                    <field name="value_ref" colspan="2" groups="base.group_advanced_attribute"/>
-                    <field name="sequence" colspan="2"/>
-                    <field name="attribute_id" colspan="2"/>
+                    <sheet>
+                        <group>
+                            <field name="name" colspan="2"/>
+                            <field name="value_ref" colspan="2" groups="base.group_advanced_attribute"/>
+                            <field name="sequence" colspan="2"/>
+                            <field name="attribute_id" colspan="2"/>
+                        </group>
+                    </sheet>
                 </form>
             </field>
         </record>
